upload modified files
