+++ .gitignore
@@ -1,3 +1,5 @@
 # folder
 
 target/
+
+context-datasource.xml
