+++ template_0709/src/main/java/com/mylibrary/book/admin/controller/HrempController.java
@@ -2,6 +2,7 @@ package com.mylibrary.book.admin.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
@@ -13,6 +14,7 @@ public class HrempController {
 	@Autowired
 	HrempService hrempService;
 	
+	// page-panning methods //
 	@RequestMapping("/hrMain")
 	public ModelAndView showMain() {
 		ModelAndView mav = new ModelAndView();
@@ -21,4 +23,17 @@ public class HrempController {
 		mav.setViewName("admin/hr-main");
 		return mav;
 	}
+	
+	@RequestMapping("/hrUpdate")
+	public String hrUpdate(Model model) {
+		return "admin/hr-update";
+	}
+	
+	@RequestMapping("/hrInsert")
+	public String hrInsert(Model model) {
+		return "admin/hr-insert";
+	}
+	// ------------------------- //
+	
+	
 }
