+++ template_0709/src/main/webapp/WEB-INF/spring/context-datasource.xml
@@ -10,7 +10,7 @@
 			value="oracle.jdbc.driver.OracleDriver" />
 		<property name="url"
 			value="jdbc:oracle:thin:@localhost:1521:XE" />
-		<property name="username" value="username" />
-		<property name="password" value="password" />
+		<property name="username" value="ssssuu" />
+		<property name="password" value="9531" />
 	</bean>
 </beans>
