+++ template_0709/src/main/webapp/WEB-INF/views/admin/components/hr-main-section.jsp
@@ -3,10 +3,10 @@
 	<main class="ttr-wrapper">
 		<div class="container-fluid">
 			<div class="db-breadcrumb">
-				<h4 class="breadcrumb-title">hr-main</h4>
+				<h4 class="breadcrumb-title">HR-Main</h4>
 				<ul class="db-breadcrumb-list">
 					<li><a href="#"><i class="fa fa-home"></i>Home</a></li>
-					<li>hr-main</li>
+					<li>HR-Main</li>
 				</ul>
 			</div>
 			<div class="row">
@@ -14,7 +14,7 @@
 				<div class="col-lg-12 m-b30">
 					<div class="widget-box">
 						<div class="wc-title">
-							<h4>hr-main</h4>
+							<h4>HR-Main</h4>
 						</div>
 						<div class="widget-inner"></div>
 						<br />
@@ -22,8 +22,11 @@
 							<div class="container-fluid">
 								<div class="card mb-4">
 									<div class="card-header">
-										<i class="fas fa-table mr-1"></i> hr-main
+										<i class="fas fa-table mr-1"></i> HR-Main
 									</div>
+									<div style="text-align:right;padding-right:20px;padding-top:20px;">
+												<button type="button" class="btn-secondry m-r5" onclick="location.href='hrInsert'">Insert</button>
+											</div>
 									<div class="card-body">
 										<div class="table-responsive">
 											<table class="table table-bordered" id="dataTable"
@@ -35,19 +38,20 @@
 														<th>Pay</th>
 														<th>Position</th>
 														<th>Department</th>
-														<th>Update</th>
+														<th>Option</th>
 														<!-- <th>Update</th> -->
 													</tr>
 												</thead>
-												<tfoot>
+												<tfoot style="border-color:white;">
 													<tr>
 														<th>Name</th>
 														<th>Tel</th>
 														<th>Pay</th>
 														<th>Position</th>
 														<th>Department</th>
-														<th>Update</th>
+														<th>Option</th>
 														<!-- <th>Update</th> -->
+														
 													</tr>
 												</tfoot>
 												<tbody>
@@ -58,8 +62,8 @@
 														<td>${itm.pay }</td>
 														<td>${itm.position }</td>
 														<td>${itm.dname }</td>
-														<td><button type="button" class="btn-secondry m-r5">Insert</button>&nbsp;
-														<button type="button" class="btn-secondry m-r5">Update</button></td>
+														<td><button type="button" class="btn-secondry m-r5" onclick="location.href='hrUpdate'">Update</button>&nbsp;
+														<button type="button" class="btn-secondry m-r5">Delete</button></td>
 													</tr>
 													</c:forEach>
 												</tbody>
